unittest_Libra-X.py is created to seperately test different functionalities of Libra-X
